open_teach/requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
asciitree==0.3.3
asttokens==3.0.0
blinker==1.9.0
blosc==1.11.1
catkin-pkg==1.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
cmeel==0.53.3
cmeel-assimp==5.3.1
cmeel-boost==1.83.0
cmeel-console-bridge==1.0.2.2
cmeel-octomap==1.9.8.2
cmeel-qhull==8.0.2.1
cmeel-tinyxml==2.6.2.3
cmeel-urdfdom==3.1.1.1
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
docutils==0.21.2
eigenpy==3.5.1
etils==1.11.0
exceptiongroup==1.2.2
executing==2.1.0
Farama-Notifications==0.0.4
fasteners==0.19
Flask==3.1.0
fonttools==4.45.1
fsspec==2024.10.0
gevent==24.11.1
glfw==2.8.0
greenlet==3.1.1
gunicorn==23.0.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h5py==3.10.0
hpp-fcl==2.4.4
hydra-core==1.3.2
idna==3.10
ikpy==3.4.2
importlib_resources==6.4.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
kiwisolver==1.4.5
llvmlite==0.43.0
lmdb==1.5.1
MarkupSafe==3.0.2
matplotlib==3.8.2
matplotlib-inline==0.1.7
mpmath==1.3.0
mujoco==3.2.6
numba==0.60.0
numcodecs==0.13.1
numpy==1.26.2
numpy-quaternion==2024.0.3
omegaconf==2.3.0
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
parso==0.8.4
pexpect==4.9.0
Pillow==10.1.0
pin==2.7.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pybullet==3.2.6
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.1.1
pyrealsense2==2.54.2.5684
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0
pyzmq==25.1.1
requests==2.32.3
robot_properties_fingers==2.0.2
rospkg==1.5.1
scipy==1.11.4
shapely==2.0.6
six==1.16.0
stack-data==0.6.3
sympy==1.13.3
tk==0.1.0
tomli==2.2.1
torch==1.12.0
torchvision==0.13.0
tqdm==4.66.1
traitlets==5.14.3
trifinger_simulation==2.0.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
xacro==1.13.3
zarr==2.18.3
zipp==3.21.0
zmq==0.0.0
zope.event==5.0
zope.interface==7.2
